extend bio.json writing for video and album metadata

+++ scrapers/freeones/videos.py
@@ -1,5 +1,5 @@
 from scrapers.scraper import Video
-from datetime import datetime
+from datetime import datetime, date
 from bs4 import BeautifulSoup
 from os import path, makedirs
 import urllib.request
@@ -12,6 +12,12 @@ class FreeOnesVideo(Video):
     base_url = "https://www.freeones.com"
     slug = ""
     target = ""
+
+    def json_serial(self,obj):
+        """JSON serializer for objects not serializable by default json code"""
+        if isinstance(obj, (datetime, date)):
+            return obj.isoformat()
+        raise TypeError ("Type %s not serializable" % type(obj))
 
     def __init__(self, slug=None, title="", target=None):
         super().__init__()
@@ -68,7 +74,7 @@ class FreeOnesVideo(Video):
 
         return requests.get(self.source_url).text
 
-    def download(self):
+    def download(self, download=False, metadata=False):
         # Load if needed
         if not self.loaded:
             self.load()
@@ -76,8 +82,16 @@ class FreeOnesVideo(Video):
         last_slug = self.slug.split('/')[-1]
         videos_path = f"./babes/{self.target}/videos"
         makedirs(videos_path, exist_ok=True)
+        json_file = f"{videos_path}/{last_slug}.json"
         target_file = f"{videos_path}/{last_slug}.mp4"
-        if not path.exists(target_file):
+
+        if metadata and not path.exists(json_file):
+            if not path.exists(json_file):
+                with open(json_file,'w') as fh:
+                    fh.write(json.dumps(self.metadata,default=self.json_serial))
+                logging.debug(f"Metadata downloaded ({json_file})")
+  
+        if download and not path.exists(target_file):
             try:
                 urllib.request.urlretrieve(self.source_url, target_file)
                 logging.info(f"Video downloaded: {self.source_url} -> {target_file}")
